+++ example/lib/main.dart
@@ -2,9 +2,9 @@ import 'dart:io';
 
 import 'package:desktop_drop/desktop_drop.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_math_fork/flutter_math.dart';
 import 'package:gpt_markdown/custom_widgets/selectable_adapter.dart';
 import 'package:gpt_markdown/gpt_markdown.dart';
-import 'package:flutter_math_fork/flutter_math.dart';
 import 'package:watcher/watcher.dart';
 
 void main() {
@@ -75,21 +75,15 @@ class MyHomePage extends StatefulWidget {
 class _MyHomePageState extends State<MyHomePage> {
   TextDirection _direction = TextDirection.ltr;
   final TextEditingController _controller = TextEditingController(
-    text: r'''
-decsiob (*) is on the set PQ = {91, 905} jjjzjsx * jjdbhsjsjmamajmsghdhhi msnnsjnskaksjjshahsh
-
-(*)
-
-This is a sample markdown document.
-
+    text: r'''This is a sample markdown document.
 * **bold**
 * *italic*
 * **_bold and italic_**
 * ~~strikethrough~~
 * `code`
-* [link](https://www.google.com) ![image](https://www.google.com/images/branding/googlelogo/1x/googlelogo_color_272x92dp.png)
+* [link](https://www.google.com)
 
-
+[![alt text](https://www.google.com/images/branding/googlelogo/1x/googlelogo_color_272x92dp.png)](link_url)
 ```markdown
 # Complex Markdown Document for Testing
 
@@ -207,21 +201,21 @@ Additional table with alignment:
 ## Blockquotes and Nested Elements
 
 > **Blockquote Header**
-> 
+>
 > This is a blockquote. You can include **bold** and *italic* text, as well as `inline code` within blockquotes.
-> 
+>
 > > ### Nested Blockquote
 > > - Nested list item 1
 > > - Nested list item 2
 > >   1. Numbered subitem 1
 > >   2. Numbered subitem 2
-> > 
+> >
 > > ```python
 > > # Code snippet inside nested blockquote
 > > for i in range(3):
 > >     print(i)
 > > ```
-> 
+>
 > Back to the outer blockquote.
 
 ---
@@ -283,11 +277,11 @@ Horizontal rules can be used to separate sections:
 
 ### Nested Quotes with Code and Math
 > **Example of Nested Components**
-> 
+>
 > - Inline code: `sum = a + b`
 > - Math expression: \( \sum_{i=1}^n i = \frac{n(n+1)}{2} \)
 > - More text with **bold** formatting.
-> 
+>
 > ```javascript
 > // JavaScript code example inside a nested blockquote
 > const sum = (n) => (n * (n + 1)) / 2;
@@ -335,6 +329,12 @@ This document was created to test the robustness of Markdown parsers and to ensu
 
   @override
   Widget build(BuildContext context) {
+//     var data = '''|asdfasfd|asdfasf|
+// |---|---|
+// |sohag|asdfasf|
+// |asdfasf|asdfasf|
+// ''';
+
     return GptMarkdownTheme(
       gptThemeData: GptMarkdownTheme.of(context).copyWith(
         highlightColor: Colors.purple,
@@ -417,6 +417,7 @@ This document was created to test the robustness of Markdown parsers and to ensu
                         ListenableBuilder(
                           listenable: _controller,
                           builder: (context, _) {
+                            var data = _controller.text;
                             return Container(
                               padding: const EdgeInsets.all(8),
                               decoration: BoxDecoration(
@@ -446,7 +447,7 @@ This document was created to test the robustness of Markdown parsers and to ensu
                                 child: Builder(
                                   builder: (context) {
                                     Widget child = GptMarkdown(
-                                      _controller.text,
+                                      data,
                                       textDirection: _direction,
                                       onLinkTap: (url, title) {
                                         debugPrint(url);
@@ -457,8 +458,9 @@ This document was created to test the robustness of Markdown parsers and to ensu
                                       textAlign: TextAlign.justify,
                                       textScaler: const TextScaler.linear(1),
                                       style: const TextStyle(
-                                        fontSize: 15,
-                                      ),
+                                          // fontFamily: 'monospace',
+                                          // fontWeight: FontWeight.bold,
+                                          ),
                                       highlightBuilder: (context, text, style) {
                                         return Container(
                                           padding: const EdgeInsets.symmetric(
@@ -520,15 +522,13 @@ This document was created to test the robustness of Markdown parsers and to ensu
                                             RegExp(r"align\*"),
                                             (match) => "aligned");
                                       },
-                                      imageBuilder:
-                                          (context, url, semanticLabel) {
-                                        return Image.network(
-                                          url,
-                                          width: 100,
-                                          height: 100,
-                                          semanticLabel: semanticLabel,
-                                        );
-                                      },
+                                      // imageBuilder: (context, url) {
+                                      //   return Image.network(
+                                      //     url,
+                                      //     width: 100,
+                                      //     height: 100,
+                                      //   );
+                                      // },
                                       latexBuilder:
                                           (context, tex, textStyle, inline) {
                                         if (tex.contains(r"\begin{tabular}")) {
@@ -614,13 +614,31 @@ This document was created to test the robustness of Markdown parsers and to ensu
                                       },
                                       linkBuilder:
                                           (context, label, path, style) {
-                                        return Text(
+                                        return Text.rich(
                                           label,
                                           style: style.copyWith(
                                             color: Colors.blue,
                                           ),
                                         );
                                       },
+
+                                      // tableBuilder: (context, tableRows,
+                                      //     textStyle, config) {
+                                      //   return Table(
+                                      //     border: TableBorder.all(
+                                      //       width: 1,
+                                      //       color: Colors.red,
+                                      //     ),
+                                      //     children: tableRows.map((e) {
+                                      //       return TableRow(
+                                      //         children: e.fields.map((e) {
+                                      //           return Text(e.data);
+                                      //         }).toList(),
+                                      //       );
+                                      //     }).toList(),
+                                      //   );
+                                      // },
+
                                       // components: [
                                       //   CodeBlockMd(),
                                       //   NewLines(),