+++ lib/custom_widgets/markdown_config.dart
@@ -44,9 +44,18 @@ typedef LatexBuilder =
 typedef LinkBuilder =
     Widget Function(
       BuildContext context,
-      String text,
+      InlineSpan text,
       String url,
       TextStyle style,
+    );
+
+/// A builder function for the table.
+typedef TableBuilder =
+    Widget Function(
+      BuildContext context,
+      List<CustomTableRow> tableRows,
+      TextStyle textStyle,
+      GptMarkdownConfig config,
     );
 
 /// A builder function for the highlight.
@@ -84,6 +93,7 @@ class GptMarkdownConfig {
     this.overflow,
     this.components,
     this.inlineComponents,
+    this.tableBuilder,
   });
 
   /// The direction of the text.
@@ -143,6 +153,9 @@ class GptMarkdownConfig {
   /// The list of inline components.
   final List<MarkdownComponent>? inlineComponents;
 
+  /// The table builder.
+  final TableBuilder? tableBuilder;
+
   /// A copy of the configuration with the specified parameters.
   GptMarkdownConfig copyWith({
     TextStyle? style,
@@ -164,6 +177,7 @@ class GptMarkdownConfig {
     final UnOrderedListBuilder? unOrderedListBuilder,
     final List<MarkdownComponent>? components,
     final List<MarkdownComponent>? inlineComponents,
+    final TableBuilder? tableBuilder,
   }) {
     return GptMarkdownConfig(
       style: style ?? this.style,
@@ -185,6 +199,7 @@ class GptMarkdownConfig {
       unOrderedListBuilder: unOrderedListBuilder ?? this.unOrderedListBuilder,
       components: components ?? this.components,
       inlineComponents: inlineComponents ?? this.inlineComponents,
+      tableBuilder: tableBuilder ?? this.tableBuilder,
     );
   }